server/authentication/seralizers.py
from rest_framework import serializers
from authentication.models import UserProfile


class UserProfileSerializer(serializers.ModelSerializer):
    class Meta:
        model = UserProfile
        fields = '__all__'