server/requirements.txt
asgiref==3.7.2
Django==4.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
dnspython==2.4.2
gunicorn==21.2.0
packaging==23.2
psycopg2==2.9.9
psycopg2-binary==2.9.9
pymongo==4.6.0
python-snappy==0.6.1
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
winkerberos==0.9.1
